фиксит бесконечный while фиксит модульность
