+++ index.html
@@ -247,7 +247,7 @@
       <button type="button" class="error__button">Попробовать снова</button>
     </div>
   </template>
-  <script src="/js/main.js" type="module"></script>
+  <script src="/js/main.mjs" type="module"></script>
 </body>
 
 </html>
