--- js/main.js
@@ -1,5 +1,0 @@
-//constants
-
-import { createObjectsArray } from './data.js';
-
-createObjectsArray();
