+++ js/main.mjs
@@ -1,0 +1,5 @@
+//constants
+
+import { createObjectsArray } from './data.mjs';
+
+createObjectsArray();
